View album details
New option to view the details of a particular album.  This is more
tooling for investigating the current state of the gallery to ensure
that we're sending up the right data.

View works, but is slow.  Just to view a single album's details, we load
all of the albums and photos in the entire gallery.  Need to put in some
on-demand loading.