Speed up viewing album details
Viewing a single album details required loading all of the albums and
all of their photo data.  This shuffles things around to significantly
speed things up.

The interface for accessing albums and photos from the GalleryFilesystem
is simplified to use dictionary lookup for access to keep testing
simple, but the actual implementation uses a variation of defaultdict to
do a lazy load on the album.