Handle subalbums as well as photos in photos.dat
Subalbums are handled similar to photos, as an extra property on the
gallery, which maps the parent album name to the list of album objects.

The GalleryFilesystem properties which perform actual file reads are all
being cached/memoized.  This should help performance now that both
albums and subalbums are reading from the same file data.